parladata/requirements.txt
Django==5.0.7
django-autoslug
psycopg==3.2.1
djangorestframework
django-filter
gunicorn
django-taggit
django-oauth-toolkit
django-extensions
ipython
raven
regex
requests
django-cors-headers
django-model-utils
django-tinymce
nltk
slugify
Pillow
pytest
pytest-django
pytest-cov
drf-spectacular
pytest-xdist
python-memcached
django-debug-toolbar
boto3==1.34.144
django-storages
pylibmc==1.6.3
martor
django-colorfield
django-admin-sortable2
pyicu
sentry-sdk[django]
eventregistry
tablib
django-import-export==4.1.1
djangorestframework-csv==3.0.2
drf-recaptcha==3.1.0
xmltodict==0.13.0

# tagger/lemmatizer
scikit-learn==1.5.1
python-crfsuite
marisa-trie
git+https://github.com/Desklop/Uk_Stemmer
